refactor: change otp route to payments router and move hander to payments controller
