+++ routes/payments/payments.router.js
@@ -1,9 +1,10 @@
 const express = require("express");
 
-const { httpGetCustomerPayments } = require("./payments.controller")
+const { httpGetCustomerPayments, httpHandleOtp } = require("./payments.controller")
 
 const paymentsRouter = express.Router();
 
 paymentsRouter.get("/:customerId", httpGetCustomerPayments);
+paymentsRouter.post("/otp", httpHandleOtp);
 
 module.exports = paymentsRouter;
